+++ libuser/src/qsort.c
@@ -96,7 +96,7 @@
 //#include <stdlib.h>
 
 static inline char	*med3(char *, char *, char *, int (*)(const void *, const void *));
-static inline void	 swapfunc(char *, char *, int, int);
+static inline void	 swapfunc(char *, char *, uint32_t, int);
 
 #define min(a, b)	(a) < (b) ? a : b
 
@@ -104,7 +104,7 @@ static inline void	 swapfunc(char *, char *, int, int);
  * Qsort routine from Bentley & McIlroy's "Engineering a Sort Function".
  */
 #define swapcode(TYPE, parmi, parmj, n) { 		\
-	long i = (n) / sizeof (TYPE); 			\
+	uint32_t i = (n) / sizeof (TYPE); 			\
 	TYPE *pi = (TYPE *) (parmi); 			\
 	TYPE *pj = (TYPE *) (parmj); 			\
 	do { 						\
@@ -114,23 +114,23 @@ static inline void	 swapfunc(char *, char *, int, int);
         } while (--i > 0);				\
 }
 
-#define SWAPINIT(a, es) swaptype = ((char *)a - (char *)0) % sizeof(long) || \
-	es % sizeof(long) ? 2 : es == sizeof(long)? 0 : 1;
+#define SWAPINIT(a, es) swaptype = ((char *)a - (char *)0) % sizeof(uint32_t) || \
+	es % sizeof(uint32_t) ? 2 : es == sizeof(uint32_t)? 0 : 1;
 
 static inline void
-swapfunc(char *a, char *b, int n, int swaptype)
+swapfunc(char *a, char *b, uint32_t n, int swaptype)
 {
 	if(swaptype <= 1) 
-		swapcode(long, a, b, n)
+		swapcode(uint32_t, a, b, n)
 	else
 		swapcode(char, a, b, n)
 }
 
 #define swap(a, b)					\
 	if (swaptype == 0) {				\
-		long t = *(long *)(a);			\
-		*(long *)(a) = *(long *)(b);		\
-		*(long *)(b) = t;			\
+		uint32_t t = *(uint32_t *)(a);			\
+		*(uint32_t *)(a) = *(uint32_t *)(b);		\
+		*(uint32_t *)(b) = t;			\
 	} else						\
 		swapfunc(a, b, es, swaptype)
